LayoutTests/platform/ios/accessibility/ignore-modals-without-any-content-expected.txt

This test ensures that we ignore modals that don't have any accessible content.

Beginning search from #container element.

StaticText
AXLabel: Foo text, before modal

StaticText
AXLabel: Foo text, after modal

Appending a text node to modal to make it accessible.
PASS: #container AX element is no longer accessible after adding accessible content to modal.

Beginning search from #modal element.

StaticText
AXLabel: Foo text, inside modal

PASS successfullyParsed is true

TEST COMPLETE
